# BUG: fix for np.unique with nan values