BUG: fix for np.unique with nan values (See https://github.com/numpy/numpy/issues/23286)

Fix np.unique which returns multiple NaNs despite setting `equal_nan=True` when `axis!=None`.

According to on-going discussion in https://github.com/numpy/numpy/issues/23286 - there should be a way to fix this behaviour using new function/ufunc. Let's treat this PR as a hot-fix.

**Brief of changes:**
- Scan over incoming input array with `(np.isnan(ar)).any()` to determine if there are any NaN values. If not, redirect function to "fallback" in if-else body.
- Keep previous optimization that was relying on `axis=None` and sorting properties. Leave comments for future devs.
- Allow dtypes of kind `V` (`np.void`) to be processed in `_unique1d()`.
- Added tests covering problematic multiple-NaN cases.
- Refactored the benchmark and parametrised it to cover 1D/2D cases with the axis set to `None`/`0`.
Usage: `python runtests.py --bench-compare main bench_lib.Unique`
- **Perf number are really close to each other and tests are passing once again.** Only significant differences are present in cases with `axis=0+` (which is currently a price for working "NaN-detection").

Main branch:
```
[ 75.00%] · For numpy commit e141d907 <main> (round 2/2):
[ 87.50%] ··· bench_lib.Unique.time_unique_1D                                                                                                                                                                    ok
[ 87.50%] ··· ============ ============= ============ ============ ============ ============= ============ ============= ============ ============= ============
              --                                                                    percent_nans / axis                                                         
              ------------ -------------------------------------------------------------------------------------------------------------------------------------
               array_size     0 / None      0 / 0      0.1 / None    0.1 / 0      2.0 / None    2.0 / 0     50.0 / None    50.0 / 0    90.0 / None    90.0 / 0  
              ============ ============= ============ ============ ============ ============= ============ ============= ============ ============= ============
                   64       3.31±0.08μs   26.0±0.8μs   3.57±0.7μs   26.2±0.6μs   3.25±0.03μs   25.7±0.6μs    5.21±0.1μs   23.7±0.4μs   5.16±0.09μs   19.8±0.2μs 
                  300       4.96±0.05μs    81.4±3μs    5.06±0.1μs    83.6±2μs     7.13±0.2μs    88.8±2μs     7.15±0.1μs    79.3±2μs     6.44±0.2μs   60.3±0.9μs 
                  512        7.32±0.2μs    161±1μs     9.28±0.5μs    151±2μs     8.85±0.08μs    146±5μs     8.11±0.02μs    127±2μs     7.42±0.09μs    92.9±1μs  
              ============ ============= ============ ============ ============ ============= ============ ============= ============ ============= ============
```

This PR:
```
[ 50.00%] · For numpy commit 7ef3abf3 <jiwaszki/np_unique_nan_bug> (round 2/2):
[ 62.50%] ··· bench_lib.Unique.time_unique_1D                                                                                                                                                                    ok
[ 62.50%] ··· ============ ============= ============ ============ ============ ============ ============ ============= ========== ============= ===========
              --                                                                  percent_nans / axis                                                       
              ------------ ---------------------------------------------------------------------------------------------------------------------------------
               array_size     0 / None      0 / 0      0.1 / None    0.1 / 0     2.0 / None    2.0 / 0     50.0 / None   50.0 / 0   90.0 / None    90.0 / 0 
              ============ ============= ============ ============ ============ ============ ============ ============= ========== ============= ===========
                   64        3.46±0.5μs    27.5±2μs    3.45±0.1μs   27.8±0.8μs   3.42±0.1μs   27.3±0.5μs    5.01±0.1μs   36.4±7μs   5.19±0.03μs    25.0±1μs 
                  300       5.15±0.06μs   83.0±0.9μs    5.57±3μs     93.4±3μs    8.02±0.3μs    104±30μs     7.26±0.8μs   86.6±3μs    6.17±0.2μs   84.2±20μs 
                  512        7.75±0.2μs    154±2μs     8.94±0.6μs    152±8μs     8.98±0.5μs    158±6μs      8.60±0.1μs   134±7μs     7.28±0.5μs    99.2±3μs 
              ============ ============= ============ ============ ============ ============ ============ ============= ========== ============= ===========
```

**TODO:**
There is another one "early return" case which I am considering. If the input array is already 1D, `axis` value is not affecting the output shape as it can be only set to either `None` or `0`, axis is swaping with itself:
```python
>>> a = np.array([0., np.nan, 2., np.nan, 2., 1., 0., 1., 2., 0.])
>>> np.unique(a, equal_nan=True, axis=None)
array([ 0.,  1.,  2., nan])
>>> a = np.array([0., np.nan, 2., np.nan, 2., 1., 0., 1., 2., 0.])
>>> np.unique(a, equal_nan=True, axis=0)
array([ 0.,  1.,  2., nan])
```