@MatteoRaso 2023-02-27
Can you add a test to see if this also works for masked arrays? If possible, it'd be nice if this PR can close #23281 as well.