@jiwaszki 2023-02-27
> Can you add a test to see if this also works for masked arrays? If possible, it'd be nice if this PR can close #23281 as well.

Sure thing! But first, regarding the problems with `axis` argument. Currently, it seems that the function is broken when combining `axis` with anything. Example:
```python
>>> import numpy as np
>>> import numpy.ma as ma
>>> a = ma.array([1,2,3,4,2,3,1,4], mask=[0,0,1,1,0,0,0,1])
>>> 
>>> out = np.unique(a, return_index=True)
>>> print(type(out), out)
<class 'tuple'> (masked_array(data=[1, 2, 3, --],
             mask=[False, False, False,  True],
       fill_value=999999), array([0, 1, 5, 2]))
>>> 
>>> out = np.unique(a, axis=0, return_index=True)
>>> print(type(out), out)
<class 'tuple'> (array([1, 2, 3, 4]), array([0, 1, 2, 3]))
```
It will take me some extra time to figure out how to fix it and keep performance "as-is".

Thanks for the heads up!