@seberg 2023-03-06
I have my doubts that this works.  The NaN can be only in one place (i.e. the first), in which case it isn't even sorted to the end (`np.unique([[np.nan, 0], [1, 2]], axis=0)`, the value of axis doesn't really matter).

It is indeed interesting whether this helps the masked array code paths, but likely the masked array fix lies not here but must be in `np.ma` itself (it is fine if `np.unique` does the wrong thing when given a masked array).